requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
cssselect==1.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.4
Django==3.1.7
django-allauth==0.44.0
django-braces==1.14.0
django-contrib-comments==2.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-generic-positions==0.2.3
django-hvad==1.8.0
django-libs==2.0.3
django-model-reviews==1.4.0
django-rest-framework==0.1.0
django-review==1.10.0
django-user-media==1.2.5
django-webpack-loader==1.0.0
djangorestframework==3.12.4
easy-thumbnails==2.7.1
gunicorn==20.1.0
idna==2.10
lxml==4.6.3
oauthlib==3.1.0
parsel==1.6.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
selectorlib==0.16.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.58.0
urllib3==1.26.4
w3lib==1.22.0
whitenoise==5.2.0
